// src/app/shared/models/installment-plan-model.ts
export class InstallmentPlanModel {
    planCode: string;
    planName: string;
    planType: string;
    loanAmt: number;
    totalTerm: number;
    interestRate: number;
    rateType: string;
    insAmt: number;

    seq: number;
    startTerm: number;
    endTerm: number;
    rate: number;
    desc: string;
}